체중 추가 기능 추가

+++ lib/db/daily_body_stats.dart
@@ -2,7 +2,7 @@ class DailyBodyStats {
   final int? id;
   final double? weight;
   final double? skeletalMuscle;
-  final int? fatRate;
+  final double? fatRate;
   final String? note;
   final String createdAt;
   final String updatedAt;
